Add support for easily specifying flexible region.
This commit also changes the meaning of --autobox_add.  This distance
is now the spacing that is added to all six sides instead of the spacing
for each of the three dimensions (e.g., the default goes from 8 to 4
but the box size stays the same).  I've also change autobox to be centered
at the middle of the bounding box of the ligand instead of the center of
mass.  The center of mass approach could potentially generate a bounding
box that doesn't fully contain the ligand if it is extremely asymmetrical.

Hopefully those changes don't screw things up too much..

The more significant change is the addition of the
--flexres
--flexdist
options which let you easily specify a flexible region of the receptor.
--flexres takes a comma separated list of chain:resid specifiers.  For example:
--flexres A:37,A:38,B:123
If the chain is omitted:
--flexres 37,38
then the first chain of the receptor is used (convenient short hand if there is only one chain).

--flexdist specifies a distance.  Any residue with a heavy atom center that is less than this
distance away from a heavy atom center of the autobox_ligand get marked as flexible.
I'm probably going change this to require a separate ligand...