implement gfloat3 constructor for float3
This was left blank, so code that used make_float3 to set float3
variables (e.g. gridmaker) was left with random values instead of an
initialized float3 (since float3 gets redefined to be gfloat3).
This meant that the molgridder was essentially
generating random values and caffe just stopped working.  Silently. And
the problem wasn't in caffe code so git searching didn't work. And on
perigee the uninitialized values often were the right values (different
toolchain) which was a total red herring.

I've never been a fan of redefining a built-in type.  Less so now.