Implement global pooling in CUDNN pooling layer
This was not implemented, which causes the gninamin test to fail (since
the overlap model uses global pooling).  The fact this has gone
unnoticed for so long is disturbing as it means we haven't been building
with CUDNN or haven't been running tests after each build.

Also snuck in a fix (hopefully) for atom typing mol2 files where we
ensure aromaticity is perceived before deleting hydrogens.