Preserve PYTHONPATH
Right now, the `$PYTHONPATH` environment variable is overwritten, and its initial value is lost, because `${PYTHONPATH}` is not defined in the `CMakeLists.txt` file. This prevents CMake from using the environment `$PYTHONPATH` and from finding Python modules (like NumPy) that may be installed in different locations.

This changes define the variable in CMake, based on the environment variable, so it can be properly extended, not overwritten.