Add Spack installation instructions to README.md
I finally had someone else reporting a successful GNINA installation with Spack, so it might be worth adding the installation instructions to the `README` for everyone to try. 

My hope is that this will facilitate GNINA adoption even further, and reduce the number of installation issues in the issue tracker.

However, I have no idea if these instructions will actually decrease the number of installation issues. There is always the risk that installation issues will increase with an additional installation method available. Hopefully they will not, and I tried to re-direct installation issues with Spack to the Spack repository.

Before merging this PR, it would be great is other people could try out the installation instructions and provide additional feedback. I'd be happy to help, and to incorporate the feedback in this PR (and the upstream Spack package).